+++ R/lme.gev11.R
@@ -49,8 +49,8 @@
 #'
 #' @examples
 #' \donttest{
-#' data(glanteifi)
-#' result <- lme.gev11(glanteifi$flow, ntry = 5)
+#' data(PhliuAgromet)
+#' result <- lme.gev11(PhliuAgromet$prec, ntry = 5)
 #' print(result$lme.gev11)
 #' }
 #'