+++ R/magev.R
@@ -98,8 +98,8 @@
 #'
 #' @examples
 #' # Load example data
-#' data(streamflow)
-#' x <- streamflow$r1
+#' data(haenam)
+#' x <- haenam$X1
 #'
 #' # Basic usage with likelihood weights
 #' result <- ma.gev(x, quant = c(0.95, 0.99), weight = 'like1', B = 100)