+++ R/strup.gev11.R
@@ -42,8 +42,8 @@
 #' @author Yonggwan Shin, Seokkap Ko, Jihong Park, Yire Shin, Jeong-Soo Park
 #'
 #' @examples
-#' data(glanteifi)
-#' result <- strup.gev11(glanteifi$flow)
+#' data(PhliuAgromet)
+#' result <- strup.gev11(PhliuAgromet$prec)
 #' print(result$strup.mdfy)
 #'
 #' @export